--- app/containers/HomeContainer/Loadable.js
@@ -1,9 +1,0 @@
-/**
- *
- * Asynchronously loads the component for HomeContainer
- *
- */
-
-import loadable from 'utils/loadable';
-
-export default loadable(() => import('./index'));
