+++ app/translations/en.json
@@ -1,14 +1,13 @@
 {
-  "repo_list": "Repository List",
-  "repo_search": "Repository Search",
-  "respo_search_default": "Search for a repository by entering it's name in the search box",
-  "get_repo_details": "Get details of repositories",
-  "search_query": "Search query: {repoName}",
-  "matching_repos": "Total number of matching repos: {totalCount}",
+  "Itunes_list": "Itunes List",
+  "artist_search": "Search Track by Artists Name",
+  "artist_search_default": "Search for a songs by entering it's artist name in the search box",
+  "track_repo_details": "Get details of tracks",
+  "search_query": "Search query: {artistName}",
+  "matching_tracks": "Total number of matching tracks: {resultCount}",
   "something_went_wrong": "Sorry. Something went wrong! Please try again in sometime.",
   "stories": "Go to Storybook",
-  "repository_name": "Repository Name: {name}",
-  "repository_full_name": "Repository full name: {fullName}",
-  "repository_stars": "Repository stars: {stars}",
+  "artist_name": "Artists Name: {artistName}",
+  "artist_full_name": "Details full name: {fullName}",
   "wednesday_solutions": "Wednesday Solutions"
 }
